Moved all tests into test_validate.py
